Split out errors into a module
This makes the docs a bit more structured.
